Исправил teplate. Перенес функцию в functions
